:truck: Standardize naming of travel request wizard steps.
Remove orphaned step component.
